% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackageSpring2021-package.R
\docType{package}
\name{testpackageSpring2021-package}
\alias{testpackageSpring2021}
\alias{testpackageSpring2021-package}
\title{testpackageSpring2021: My Package as a Demo Package}
\description{
This is my basic package.
}
\author{
\strong{Maintainer}: Rebecca Lipton \email{rebeccarl013@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
